+++ src/pages/accounts/components/index.ts
@@ -1,5 +1,8 @@
+export { DashboardProgress } from "./DashboardProgress";
 export { AccountsTable } from "./AccountsTable";
 export { NewAccountDialog } from "./NewAccountDialog";
-export { DashboardProgress } from "./DashboardProgress";
 export { AccountDialog } from "./AccountDialog";
+export { AccountInfo } from "./AccountInfo";
+export { AdminActions } from "./AdminActions";
+export { AccountEditForm } from "./AccountInfoEdit";
 export { AccountSnackbar } from "./AccountSnackbar";
